Use new dynamic shader system
WIth this system shader reflection is used to get the input layout from the vertex shader, then when the mesh is binded to a shader it automatically sends only the data that the shader needs. If a model does not have the data that a shader needs there should be a helpful assert message.

This should make it easier to add more shader types in the future.