Get normal mapping working on car scene
To get this working I had to do a pretty ugly hack. The different meshes of the car have to be rendered differently since some of them don't have normal maps. To do this I made the BumpMap pipeline fall back to the LightPipeline if there is no normal map. It'll have to do until I figure out how I want this to work.