Remove IGraphics interface
Only reason I added it to begin with was to quickly get rid of a circular dependency, but that's no longer an issue now and I can remove the interface without the clutter. To aid in removing the dependency mess I added a new RenderContext class that lets graphics modules get the device/context without depending on Graphics.